feat: decouple dashboard statistics from recent activity feed
Decoupled Total Matches, Win %, and Current Streak from the limited "Recent Activity" query.
Lifetime statistics are now calculated by querying all historical matches for the user, while the activity feed remains limited to the latest 20 matches for performance.

Key changes:
- Updated `UserService.get_dashboard_data` and `_fetch_vanity_stats` to calculate true lifetime metrics.
- Updated `UserService.get_user_profile_data` and `_fetch_profile_stats` to use all matches for lifetime stats.
- Modified `pickaladder/user/routes.py` to pass decoupled lifetime variables to the profile template.
- Updated `pickaladder/api/stats_routes.py` to ensure lazy-loaded vanity metrics reflect accurate historical data.
- Enhanced `pickaladder/templates/user/profile.html` to display lifetime match totals and use correct streak data.
- Fixed linting and formatting issues in service files.
- Verified accurate calculation and display of lifetime stats via unit tests.
