+++ pickaladder/match/services/command.py
@@ -17,7 +17,6 @@ if TYPE_CHECKING:
     from google.cloud.firestore_v1.batch import WriteBatch
     from google.cloud.firestore_v1.client import Client
     from google.cloud.firestore_v1.document import DocumentReference
-    from google.cloud.firestore_v1.transaction import Transaction
 
     from pickaladder.user.models import UserSession
 
@@ -35,8 +34,6 @@ class MatchCommandService(BaseRepository):
         current_user: UserSession,
     ) -> MatchResult:
         """Process and record a match submission."""
-        from firebase_admin import firestore
-
         user_id = current_user["uid"]
         sub = data if isinstance(data, MatchSubmission) else MatchSubmission(**data)
         MatchValidationService.validate_submission(db, sub, user_id)
@@ -49,30 +46,18 @@ class MatchCommandService(BaseRepository):
         new_match_ref = cast(
             "DocumentReference", db.collection(cls.COLLECTION_NAME).document()
         )
-        transaction = db.transaction()
-
-        @firestore.transactional
-        def _execute_match_transaction(
-            transaction: Transaction,
-            match_ref: DocumentReference,
-            side_refs: tuple[DocumentReference, DocumentReference],
-            user_ref: DocumentReference,
-            match_data: dict[str, Any],
-        ) -> None:
-            cls._record_match_atomic(
-                db,
-                transaction,
-                (match_ref, side_refs, user_ref),
-                match_data,
-            )
-
-        _execute_match_transaction(
-            transaction,
+        batch = db.batch()
+        cls._record_match_batch(
+            db,
+            batch,
             new_match_ref,
-            (side1_ref, side2_ref),
+            side1_ref,
+            side2_ref,
             cast("DocumentReference", db.collection("users").document(user_id)),
             match_doc_data,
+            sub.match_type,
         )
+        batch.commit()
 
         return cls._build_match_result(new_match_ref.id, match_doc_data)
 
@@ -154,49 +139,24 @@ class MatchCommandService(BaseRepository):
         )
 
     @classmethod
-    def _record_match_atomic(
+    def _record_match_batch(  # noqa: PLR0913
         cls,
         db: Client,
-        transaction: Transaction | WriteBatch,
-        refs: tuple[
-            DocumentReference,
-            tuple[DocumentReference, DocumentReference],
-            DocumentReference,
-        ],
+        batch: WriteBatch,
+        match_ref: DocumentReference,
+        p1_ref: DocumentReference,
+        p2_ref: DocumentReference,
+        user_ref: DocumentReference,
         match_data: dict[str, Any],
+        match_type: str,
     ) -> None:
-        """Record a match and update stats atomically."""
+        """Record a match and update stats using batched writes."""
         from firebase_admin import firestore
-        from google.cloud.firestore_v1.transaction import Transaction
 
-        match_ref, side_refs, user_ref = refs
-        p1_ref, p2_ref = side_refs
-        match_type = match_data.get("matchType", "singles")
-
-        # Perform all reads first
-        read_refs = [p1_ref, p2_ref]
-        side1_ids, side2_ids = cls._get_side_ids(match_data, match_type)
-        all_participant_ids = list(set(side1_ids + side2_ids))
-        stats_refs = []
-        if isinstance(transaction, Transaction):
-            for uid in all_participant_ids:
-                stats_refs.append(
-                    db.collection("users")
-                    .document(uid)
-                    .collection("stats")
-                    .document("lifetime")
-                )
-            read_refs.extend(stats_refs)
-
-        # get_all only accepts transaction if it's not None
-        actual_transaction = (
-            transaction if isinstance(transaction, Transaction) else None
-        )
-        snaps_list = db.get_all(read_refs, transaction=actual_transaction)
-        snaps = {s.reference.path: s for s in snaps_list if s.exists}
-
-        p1_snap = cast("DocumentSnapshot", snaps.get(p1_ref.path))
-        p2_snap = cast("DocumentSnapshot", snaps.get(p2_ref.path))
+        snaps_list = db.get_all([p1_ref, p2_ref])
+        snaps = {s.id: s for s in snaps_list if s.exists}
+        p1_snap = cast("DocumentSnapshot", snaps.get(p1_ref.id))
+        p2_snap = cast("DocumentSnapshot", snaps.get(p2_ref.id))
         p1_data = p1_snap.to_dict() if p1_snap else {}
         p2_data = p2_snap.to_dict() if p2_snap else {}
 
@@ -205,6 +165,7 @@ class MatchCommandService(BaseRepository):
                 match_data, p1_ref, p1_data or {}, p2_ref, p2_data or {}
             )
 
+        side1_ids, side2_ids = cls._get_side_ids(match_data, match_type)
         outcome = MatchStatsCalculator.calculate_match_outcome(
             match_data["player1Score"],
             match_data["player2Score"],
@@ -223,45 +184,22 @@ class MatchCommandService(BaseRepository):
         ):
             match_data["is_upset"] = True
 
-        # Now perform all writes
-        transaction.set(match_ref, match_data)
-        transaction.update(p1_ref, p1_upd)
-        transaction.update(p2_ref, p2_upd)
+        batch.set(match_ref, match_data)
+        batch.update(p1_ref, p1_upd)
+        batch.update(p2_ref, p2_upd)
 
         if match_type == "doubles":
             MatchStatsUpdater.update_user_stats_batch(
-                transaction,
+                batch,
                 match_data.get("team1", []),
                 match_data.get("team2", []),
                 outcome["winner"],
             )
 
-        if isinstance(transaction, Transaction):
-            for uid in outcome.get("winners", []):
-                s_ref = (
-                    db.collection("users")
-                    .document(uid)
-                    .collection("stats")
-                    .document("lifetime")
-                )
-                MatchStatsUpdater.update_lifetime_stats_atomic(
-                    transaction, uid, True, db, snapshot=snaps.get(s_ref.path)
-                )
-            for uid in outcome.get("losers", []):
-                s_ref = (
-                    db.collection("users")
-                    .document(uid)
-                    .collection("stats")
-                    .document("lifetime")
-                )
-                MatchStatsUpdater.update_lifetime_stats_atomic(
-                    transaction, uid, False, db, snapshot=snaps.get(s_ref.path)
-                )
-
-        transaction.update(user_ref, {"lastMatchRecordedType": match_type})
+        batch.update(user_ref, {"lastMatchRecordedType": match_type})
 
         if gid := match_data.get("groupId"):
-            transaction.update(
+            batch.update(
                 db.collection("groups").document(gid),
                 {"updatedAt": firestore.SERVER_TIMESTAMP},
             )
