+++ pickaladder/match/services/match_stats_updater.py
@@ -5,17 +5,14 @@ from typing import TYPE_CHECKING, Any
 from firebase_admin import firestore
 
 if TYPE_CHECKING:
-    from google.cloud.firestore_v1.base_document import DocumentSnapshot
     from google.cloud.firestore_v1.batch import WriteBatch
-    from google.cloud.firestore_v1.client import Client
     from google.cloud.firestore_v1.document import DocumentReference
-    from google.cloud.firestore_v1.transaction import Transaction
 
 
 class MatchStatsUpdater:
     @staticmethod
     def update_user_stats_batch(
-        batch: WriteBatch | Transaction,
+        batch: WriteBatch,
         team1_refs: list[DocumentReference],
         team2_refs: list[DocumentReference],
         winner: str,
@@ -58,49 +55,3 @@ class MatchStatsUpdater:
         if r2:
             field = "stats.wins" if not s1_won else "stats.losses"
             r2.update({field: firestore.Increment(delta)})
-
-    @staticmethod
-    def update_lifetime_stats_atomic(
-        transaction: Transaction,
-        user_id: str,
-        is_win: bool,
-        db: Client,
-        snapshot: DocumentSnapshot | None = None,
-    ) -> None:
-        """Update lifetime stats for a user atomically.
-
-        Args:
-            transaction: The transaction object.
-            user_id: The user ID.
-            is_win: Whether the match was a win.
-            db: The Firestore client.
-            snapshot: Optional pre-fetched snapshot for the lifetime stats document.
-        """
-        stats_ref = (
-            db.collection("users")
-            .document(user_id)
-            .collection("stats")
-            .document("lifetime")
-        )
-
-        if snapshot is None:
-            snapshot = stats_ref.get(transaction=transaction)
-
-        snapshot_dict = (
-            snapshot.to_dict() if (snapshot and snapshot.exists) else None
-        ) or {"total_matches": 0, "wins": 0, "losses": 0, "current_streak": 0}
-
-        # Create a mutable copy and ensure it's a dict for mypy
-        stats = dict(snapshot_dict)
-
-        stats["total_matches"] = int(stats.get("total_matches", 0)) + 1
-        streak = int(stats.get("current_streak", 0))
-
-        if is_win:
-            stats["wins"] = int(stats.get("wins", 0)) + 1
-            stats["current_streak"] = streak + 1 if streak > 0 else 1
-        else:
-            stats["losses"] = int(stats.get("losses", 0)) + 1
-            stats["current_streak"] = streak - 1 if streak < 0 else -1
-
-        transaction.set(stats_ref, stats, merge=True)
