+++ pickaladder/static/css/cards.css
@@ -102,6 +102,7 @@
     position: relative;
     max-width: 350px;
     width: fit-content;
+    min-width: 280px;
     padding: 16px 8px;
     display: flex;
     flex-direction: column;
@@ -274,8 +275,6 @@
 .tournament-card {
     height: 100%;
     width: 100%;
-    max-width: 100%;
-    min-width: 0;
     overflow: hidden;
 }
